pyproject.toml
[project]
name = "zndraw"
version = "0.6.0a4"
description = "ZnDraw"
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=6.2.0",
    "celery>=5.5.3",
    "eventlet>=0.40.3",
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "msgpack>=1.1.1",
    "networkx>=3.5",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pydantic>=2.11.9",
    "pyjwt>=2.10.1",
    "python-socketio[client]>=5.13.0",
    "rdkit2ase>=0.1.13",
    "redis>=6.4.0",
    "tqdm>=4.67.1",
    "typer>=0.19.2",
    "vesin>=0.3.8",
    "zarr>=3.1.3",
    "znh5md>=0.4.8",
    "znsocket>=0.2.13a3",
]

[dependency-groups]
dev = [
    "fastmcp>=2.12.4",
    "ipykernel>=6.30.1",
    "nbformat>=5.10.4",
    "pandas>=2.3.3",
    "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zndraw"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/*.pyc",
    "**/.DS_Store",
    "app/node_modules",
    "dist",
    ".git",
    ".vscode",
    ".claude",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/*.pyc",
    "**/.DS_Store",
    "app/node_modules",
    "dist",
    ".git",
    ".vscode",
    ".claude",
]

[project.scripts]
zndraw = "zndraw.cli:app"

[project.optional-dependencies]
test = [
    "pytest>=8.4.2",
]
